Handle failed sign-in
This ensures that the effect fails on sign-in fail. This has to be
derived from the body of a successful HTTP response due to how
MBO works, so we look for a failure string.

The other problem is how STTP treats redirects. It doesn't seem to
add cookies from the original request to the redirect request so
we add that handling ourselves. The reason this is needed is
because MBO redirects to another page with the failure response
for the sign-in based on cookie.