Move conversion of readable class string
The conversion to a readable string is booking specific
because we want to also confirm the status of the booking.
If used outside of this context, the status indicator doesn't
make much sense in the case that the class isn't bookable.

In our case we know that we've booked the class so it must be
bookable so reporting the status makes sense and can inform
of an invalid status accordingly.